extract fetch logic
